Tarannis 1.2 API Changes:  * Added support for the future 64-bit Item IDs.  * Other IDs are not affected as we were already using int rather than short or byte.  * Changed more than strictly needed as it was worth avoiding having to explicitly cast between int and long.

+++ EVEMon.Common/Data/SolarSystemRange.cs
@@ -48,7 +48,7 @@ namespace EVEMon.Common.Data
         /// <returns></returns>
         internal static List<SolarSystemRange> GetSystemRangesFrom(SolarSystem system, int maxInclusiveNumberOfJumps)
         {
-            var collectedSystems = new Dictionary<int, SolarSystem>();
+            var collectedSystems = new Dictionary<long, SolarSystem>();
             var ranges = new List<SolarSystemRange>();
             var lastRange = new SolarSystemRange(system, 0);
 
@@ -68,7 +68,7 @@ namespace EVEMon.Common.Data
         /// gets the next 
         /// </summary>
         /// <returns></returns>
-        private SolarSystemRange GetNextRange(Dictionary<int, SolarSystem> collectedSystems)
+        private SolarSystemRange GetNextRange(Dictionary<long, SolarSystem> collectedSystems)
         {
             SolarSystemRange nextRange = new SolarSystemRange(m_source, m_range + 1);
